fix #168 	- I'm pretty sure this is what it is supposed to be 	- I would suggest to rename the function to `resolve_scad_filename`
